+++ com.unity.ml-agents/Runtime/Inference/BarracudaModelExtensions.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Unity.Barracuda;
+using FailedCheck = Unity.MLAgents.Inference.BarracudaModelParamLoader.FailedCheck;
 
 namespace Unity.MLAgents.Inference
 {
@@ -36,6 +37,18 @@ namespace Unity.MLAgents.Inference
             names.Sort();
 
             return names.ToArray();
+        }
+
+        /// <summary>
+        /// Get the version of the model.
+        /// </summary>
+        /// <param name="model">
+        /// The Barracuda engine model for loading static parameters.
+        /// </param>
+        /// <returns>The api version of the model</returns>
+        public static int GetVersion(this Model model)
+        {
+            return (int)model.GetTensorByName(TensorNames.VersionNumber)[0];
         }
 
         /// <summary>
@@ -226,12 +239,20 @@ namespace Unity.MLAgents.Inference
             else
             {
                 return model.outputs.Contains(TensorNames.DiscreteActionOutput) &&
-                    (int)model.GetTensorByName(TensorNames.DiscreteActionOutputShape)[0] > 0;
+                    (int)model.DiscreteOutputSize() > 0;
             }
         }
 
         /// <summary>
         /// Discrete action output size of the model. This is equal to the sum of the branch sizes.
+        /// This method gets the tensor representing the list of branch size and returns the
+        /// sum of all the elements in the Tensor.
+        ///  - In version 1.X this tensor contains a single number, the sum of all branch
+        /// size values.
+        ///  - In version 2.X this tensor contains a 1D Tensor with each element corresponding
+        /// to a branch size.
+        /// Since this method does the sum of all elements in the tensor, the output
+        /// will be the same on both 1.X and 2.X.
         /// </summary>
         /// <param name="model">
         /// The Barracuda engine model for loading static parameters.
@@ -249,7 +270,19 @@ namespace Unity.MLAgents.Inference
             else
             {
                 var discreteOutputShape = model.GetTensorByName(TensorNames.DiscreteActionOutputShape);
-                return discreteOutputShape == null ? 0 : (int)discreteOutputShape[0];
+                if (discreteOutputShape == null)
+                {
+                    return 0;
+                }
+                else
+                {
+                    int result = 0;
+                    for (int i = 0; i < discreteOutputShape.length; i++)
+                    {
+                        result += (int)discreteOutputShape[i];
+                    }
+                    return result;
+                }
             }
         }
 
@@ -298,13 +331,15 @@ namespace Unity.MLAgents.Inference
         /// <param name="failedModelChecks">Output list of failure messages</param>
         ///
         /// <returns>True if the model contains all the expected tensors.</returns>
-        public static bool CheckExpectedTensors(this Model model, List<string> failedModelChecks)
+        public static bool CheckExpectedTensors(this Model model, List<FailedCheck> failedModelChecks)
         {
             // Check the presence of model version
             var modelApiVersionTensor = model.GetTensorByName(TensorNames.VersionNumber);
             if (modelApiVersionTensor == null)
             {
-                failedModelChecks.Add($"Required constant \"{TensorNames.VersionNumber}\" was not found in the model file.");
+                failedModelChecks.Add(
+                    FailedCheck.Warning($"Required constant \"{TensorNames.VersionNumber}\" was not found in the model file.")
+                    );
                 return false;
             }
 
@@ -312,7 +347,9 @@ namespace Unity.MLAgents.Inference
             var memorySizeTensor = model.GetTensorByName(TensorNames.MemorySize);
             if (memorySizeTensor == null)
             {
-                failedModelChecks.Add($"Required constant \"{TensorNames.MemorySize}\" was not found in the model file.");
+                failedModelChecks.Add(
+                    FailedCheck.Warning($"Required constant \"{TensorNames.MemorySize}\" was not found in the model file.")
+                    );
                 return false;
             }
 
@@ -321,7 +358,9 @@ namespace Unity.MLAgents.Inference
                 !model.outputs.Contains(TensorNames.ContinuousActionOutput) &&
                 !model.outputs.Contains(TensorNames.DiscreteActionOutput))
             {
-                failedModelChecks.Add("The model does not contain any Action Output Node.");
+                failedModelChecks.Add(
+                    FailedCheck.Warning("The model does not contain any Action Output Node.")
+                    );
                 return false;
             }
 
@@ -330,13 +369,18 @@ namespace Unity.MLAgents.Inference
             {
                 if (model.GetTensorByName(TensorNames.ActionOutputShapeDeprecated) == null)
                 {
-                    failedModelChecks.Add("The model does not contain any Action Output Shape Node.");
+                    failedModelChecks.Add(
+                        FailedCheck.Warning("The model does not contain any Action Output Shape Node.")
+                        );
                     return false;
                 }
                 if (model.GetTensorByName(TensorNames.IsContinuousControlDeprecated) == null)
                 {
-                    failedModelChecks.Add($"Required constant \"{TensorNames.IsContinuousControlDeprecated}\" was not found in the model file. " +
-                        "This is only required for model that uses a deprecated model format.");
+                    failedModelChecks.Add(
+                        FailedCheck.Warning($"Required constant \"{TensorNames.IsContinuousControlDeprecated}\" was " +
+                        "not found in the model file. " +
+                        "This is only required for model that uses a deprecated model format.")
+                        );
                     return false;
                 }
             }
@@ -345,13 +389,17 @@ namespace Unity.MLAgents.Inference
                 if (model.outputs.Contains(TensorNames.ContinuousActionOutput) &&
                     model.GetTensorByName(TensorNames.ContinuousActionOutputShape) == null)
                 {
-                    failedModelChecks.Add("The model uses continuous action but does not contain Continuous Action Output Shape Node.");
+                    failedModelChecks.Add(
+                        FailedCheck.Warning("The model uses continuous action but does not contain Continuous Action Output Shape Node.")
+                        );
                     return false;
                 }
                 if (model.outputs.Contains(TensorNames.DiscreteActionOutput) &&
                     model.GetTensorByName(TensorNames.DiscreteActionOutputShape) == null)
                 {
-                    failedModelChecks.Add("The model uses discrete action but does not contain Discrete Action Output Shape Node.");
+                    failedModelChecks.Add(
+                        FailedCheck.Warning("The model uses discrete action but does not contain Discrete Action Output Shape Node.")
+                        );
                     return false;
                 }
             }