+++ com.unity.ml-agents/Runtime/Inference/TensorGenerator.cs
@@ -35,6 +35,7 @@ namespace Unity.MLAgents.Inference
         }
 
         readonly Dictionary<string, IGenerator> m_Dict = new Dictionary<string, IGenerator>();
+        int m_ApiVersion;
 
         /// <summary>
         /// Returns a new TensorGenerators object.
@@ -55,6 +56,8 @@ namespace Unity.MLAgents.Inference
                 return;
             }
             var model = (Model)barracudaModel;
+
+            m_ApiVersion = model.GetVersion();
 
             // Generator for Inputs
             m_Dict[TensorNames.BatchSizePlaceholder] =
@@ -145,34 +148,14 @@ namespace Unity.MLAgents.Inference
                 ObservationGenerator obsGen = null;
                 string obsGenName = null;
                 switch (rank)
+                for (var sensorIndex = 0; sensorIndex < sensors.Count; sensorIndex++)
                 {
-                    case 1:
-                        if (vecObsGen == null)
-                        {
-                            vecObsGen = new ObservationGenerator(allocator);
-                        }
-                        obsGen = vecObsGen;
-                        obsGenName = TensorNames.VectorObservationPlaceholder;
-                        break;
-                    case 2:
-                        // If the tensor is of rank 2, we use the index of the sensor
-                        // to create the name
-                        obsGen = new ObservationGenerator(allocator);
-                        obsGenName = TensorNames.GetObservationName(sensorIndex);
-                        break;
-                    case 3:
-                        // If the tensor is of rank 3, we use the "visual observation
-                        // index", which only counts the rank 3 sensors
-                        obsGen = new ObservationGenerator(allocator);
-                        obsGenName = TensorNames.GetVisualObservationName(visIndex);
-                        visIndex++;
-                        break;
-                    default:
-                        throw new UnityAgentsException(
-                            $"Sensor {sensor.GetName()} have an invalid rank {rank}");
+                    var obsGen = new ObservationGenerator(allocator);
+                    var obsGenName = TensorNames.GetObservationName(sensorIndex);
+                    obsGen.AddSensorIndex(sensorIndex);
+                    m_Dict[obsGenName] = obsGen;
+
                 }
-                obsGen.AddSensorIndex(sensorIndex);
-                m_Dict[obsGenName] = obsGen;
             }
         }